refactor: rename activity schema and router files - Rename app/schemas/activity.py to activity_schema.py - Rename app/api/activity.py to activity_router.py - Update all import references in main.py, activity_service.py, and activity_router.py - Improve naming consistency across the codebase
