+++ app/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI, status
 from fastapi.responses import JSONResponse
 
-from app.api.activity import router as activity_router
+from app.api.activity_router import router as activity_router
 from app.api.health import router as health_router
 from app.api.program import router as program_router
 from app.api.user_router import router as user_router
